Move button component to different file
