+++ src/components/Calculator.js
@@ -1,28 +1,6 @@
 import React, { useState } from 'react';
-import PropTypes from 'prop-types';
+import { DigitButton, OperatorButton } from './Buttons';
 import Calculate from '../logic/calculate';
-
-// DigitButton component with prop validation
-function DigitButton({ digit, onClick }) {
-  return (
-    <button type="button" className="digit-button" onClick={() => onClick(digit)}>{ digit }</button>
-  );
-}
-DigitButton.propTypes = {
-  digit: PropTypes.oneOfType([PropTypes.string, PropTypes.number]).isRequired,
-  onClick: PropTypes.func.isRequired,
-};
-
-// OpertorButton component with prop validation
-function OperatorButton({ operator, onClick }) {
-  return (
-    <button type="button" className="operator-button" onClick={() => onClick(operator)}>{ operator }</button>
-  );
-}
-OperatorButton.propTypes = {
-  operator: PropTypes.string.isRequired,
-  onClick: PropTypes.func.isRequired,
-};
 
 // Calculator App
 function Calculator() {
